Fix tokio::Bytes bench and add two alternative tokio::BytesMut benches
1. The existing bench was using `with_capacity` and then writing directly to the
BytesMut. Unfortunately this results in copying nothing into the buffer as
BytesMut impl of Deref<[u8]> only returns a slice from `0..len`.

2. Added two alternative ways to use tokio::BytesMut without needing to zero out
the buffer. The first uses `BytesMut::writer` with a performance improvement of
roughly 20% for long values (and a 20% regression on short values). The second uses
the unsafe function `BytesMut::set_len` with a performance improvement of
roughly 40% on long values and ~35% on short values.

This change also returns values from Criterion bench fns which I believe is a
slightly more idiomatic way to use Criterion and may also allow Criterion to
ensure the compiler doesn't optimize the result away.